Delays for message receive and delay retries in case of failure.
Without, SDS011 tends to either crash (red LED stays on indefinately and sensor freezes) or the stop command is not observed for a whole interval, leading to laser and fan wear out.

+++ Sds011.h
@@ -48,7 +48,7 @@ protected:
 	};
 
 	void _send_cmd(enum Command cmd, const uint8_t* buf, uint8_t len);
-	uint8_t _read_byte(long unsigned deadline = 0);
+	int _read_byte(long unsigned deadline = 0);
 	String _buf_to_string();
 	void _clear_responses();
 	bool _read_response(enum Command cmd);
